fix: Address Copilot PR review comments (#15)
Addresses all 14 code review comments from GitHub Copilot on PR #15:

1. CSP Issues (4 comments)
   - Keep 'unsafe-inline' in showFaq/showRiscvGuide/showStepZero
   - Prevents breaking template's inline scripts and event handlers
   - Template remains functional with back-to-top and smooth scrolling

2. Sudoers Syntax (3 comments)
   - Fix NOPASSWD:ALL → NOPASSWD: ALL (add required space)
   - Use /etc/sudoers.d/coder instead of appending to /etc/sudoers
   - Set proper permissions (0440) on sudoers file
   - Applies to: Dockerfile, Dockerfile.full, Dockerfile.koyeb

3. Install Script Security (4 comments)
   - Add security notice comments for code-server install
   - Add security notice for tt-installer install
   - Document trade-off between security and maintainability
   - Note: Keeping unverified installs for maintainability

4. Code Duplication (1 comment)
   - Extract getNonce() to shared src/utils/webview.ts
   - Remove duplicate from extension.ts and LessonWebviewManager.ts
   - DRY principle - single source of truth for nonce generation

5. Reproducibility (1 comment)
   - Pin tt-metal clone to v0.65.1 in Dockerfile.koyeb
   - Use --depth 1 for faster, smaller clone
   - Ensures reproducible builds

6. Documentation (1 comment)
   - Remove incorrect README.md claim from CHANGELOG.md
   - README.md was not actually modified in this PR

All changes maintain existing functionality while improving code quality,
security documentation, and build reproducibility.
