+++ Dockerfile
@@ -29,12 +29,16 @@ RUN apt-get update && apt-get install -y --no-install-recommends \
     && rm -rf /var/lib/apt/lists/*
 
 # Install code-server using official installation script
+# Security Note: This uses the official code-server install script from a trusted source.
+# While version pinning would be more secure, it adds maintenance burden.
+# Consider pinning to specific versions in production environments.
 RUN curl -fsSL https://code-server.dev/install.sh | sh
 
 # Create coder user with home directory and sudo privileges
 RUN useradd -m -s /bin/bash coder \
     && usermod -aG sudo coder \
-    && echo "coder ALL=(ALL) NOPASSWD:ALL" >> /etc/sudoers
+    && echo "coder ALL=(ALL) NOPASSWD: ALL" > /etc/sudoers.d/coder \
+    && chmod 0440 /etc/sudoers.d/coder
 
 # Install HuggingFace CLI and matplotlib for cookbook examples
 # Note: --break-system-packages is safe in containers (PEP 668)
